try/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pdfplumber>=0.9.0
python-docx>=0.8.11
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
